EP0057 - Moving the game code into a separate DLL https://www.youtube.com/watch?v=GoBHbSVwiDM

+++ ExitYesNoScreen.c
@@ -53,14 +53,14 @@ void DrawExitYesNoScreen(void)
 		g_input_enabled = TRUE;
 	}
 
-	BlitStringToBuffer(gMenu_ExitYesNo.Name, &g_6x7_font, &text_color, (GAME_RES_WIDTH / 2) - (uint16_t)(strlen(gMenu_ExitYesNo.Name) * 6 / 2), 60);
+	BlitStringToBuffer(gMenu_ExitYesNo.Name, &g_6x7_font, &text_color, (int16_t)((GAME_RES_WIDTH / 2) - (uint16_t)(strlen(gMenu_ExitYesNo.Name) * 6 / 2)), 60);
 	BlitStringToBuffer(gMenu_ExitYesNo.Items[0]->Name, &g_6x7_font, &text_color, gMenu_ExitYesNo.Items[0]->X, gMenu_ExitYesNo.Items[0]->Y);
 	BlitStringToBuffer(gMenu_ExitYesNo.Items[1]->Name, &g_6x7_font, &text_color, gMenu_ExitYesNo.Items[1]->X, gMenu_ExitYesNo.Items[1]->Y);
 
 	BlitStringToBuffer("\xBB",
 		&g_6x7_font,
 		&text_color,
-		gMenu_ExitYesNo.Items[gMenu_ExitYesNo.SelectedItem]->X - 6,
+		(int16_t)(gMenu_ExitYesNo.Items[gMenu_ExitYesNo.SelectedItem]->X - 6),
 		gMenu_ExitYesNo.Items[gMenu_ExitYesNo.SelectedItem]->Y);
 
 	local_frame_counter++;
